refactor: 패키지 이름 수정
- activity 에서 view로 변경
