+++ app/src/main/java/woowacourse/movie/view/movielist/MainActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.activity.movielist
+package woowacourse.movie.view.movielist
 
 import android.content.Intent
 import android.os.Bundle
